Use size_t instead of uint for qHash
In Qt6 the hash type is now potentially using 64bit instead
of being limited to 32bit, let us follow suit.

Fixes: https://invent.kde.org/kdevelop/kdevelop/-/issues/50

+++ kdevplatform/language/duchain/declarationid.h
@@ -190,7 +190,7 @@ private:
     IndexedInstantiationInformation m_specialization;
 };
 
-inline uint qHash(const KDevelop::DeclarationId& id)
+inline size_t qHash(const KDevelop::DeclarationId& id)
 {
     return id.hash();
 }
